move reformatting of summary data out of EventSummarizer
