+++ test/LaunchDarkly.Tests/EventSummarizerTest.cs
@@ -1,7 +1,8 @@
 ﻿using System.Collections.Generic;
-using LaunchDarkly.Client;
+using System.Linq;
 using Newtonsoft.Json.Linq;
 using Xunit;
+using LaunchDarkly.Client;
 
 namespace LaunchDarkly.Tests
 {
@@ -14,9 +15,9 @@ namespace LaunchDarkly.Tests
         public void SummarizeEventDoesNothingForIdentifyEvent()
         {
             EventSummarizer es = new EventSummarizer();
-            SummaryState snapshot = es.Snapshot();
+            EventSummary snapshot = es.Snapshot();
             es.SummarizeEvent(_eventFactory.NewIdentifyEvent(_user));
-            SummaryState snapshot2 = es.Snapshot();
+            EventSummary snapshot2 = es.Snapshot();
             Assert.Equal(snapshot.StartDate, snapshot2.StartDate);
             Assert.Equal(snapshot.EndDate, snapshot2.EndDate);
             Assert.Equal(snapshot.Counters, snapshot2.Counters);
@@ -26,9 +27,9 @@ namespace LaunchDarkly.Tests
         public void SummarizeEventDoesNothingForCustomEvent()
         {
             EventSummarizer es = new EventSummarizer();
-            SummaryState snapshot = es.Snapshot();
+            EventSummary snapshot = es.Snapshot();
             es.SummarizeEvent(_eventFactory.NewCustomEvent("whatever", _user, null));
-            SummaryState snapshot2 = es.Snapshot();
+            EventSummary snapshot2 = es.Snapshot();
             Assert.Equal(snapshot.StartDate, snapshot2.StartDate);
             Assert.Equal(snapshot.EndDate, snapshot2.EndDate);
             Assert.Equal(snapshot.Counters, snapshot2.Counters);
@@ -48,7 +49,7 @@ namespace LaunchDarkly.Tests
             es.SummarizeEvent(event1);
             es.SummarizeEvent(event2);
             es.SummarizeEvent(event3);
-            SummaryOutput data = es.Output(es.Snapshot());
+            EventSummary data = es.Snapshot();
 
             Assert.Equal(1000, data.StartDate);
             Assert.Equal(2000, data.EndDate);
@@ -78,26 +79,17 @@ namespace LaunchDarkly.Tests
             es.SummarizeEvent(event3);
             es.SummarizeEvent(event4);
             es.SummarizeEvent(event5);
-            SummaryOutput data = es.Output(es.Snapshot());
+            EventSummary data = es.Snapshot();
 
-            data.Features[flag1.Key].Counters.Sort((a, b) => ((string)a.Value).CompareTo((string)b.Value));
-            EventSummaryCounter expected1 = new EventSummaryCounter(new JValue("value1"),
-                flag1.Version, 2);
-            EventSummaryCounter expected2 = new EventSummaryCounter(new JValue("value2"),
-                flag1.Version, 1);
-            EventSummaryCounter expected3 = new EventSummaryCounter(new JValue("value99"),
-                flag2.Version, 1);
-            EventSummaryCounter expected4 = new EventSummaryCounter(default3,
-                null, 1);
-            Assert.Equal(new EventSummaryFlag(default1,
-                new List<EventSummaryCounter> { expected1, expected2 }),
-                data.Features[flag1.Key]);
-            Assert.Equal(new EventSummaryFlag(default2,
-                new List<EventSummaryCounter> { expected3 }),
-                data.Features[flag2.Key]);
-            Assert.Equal(new EventSummaryFlag(default3,
-                new List<EventSummaryCounter> { expected4 }),
-                data.Features[unknownFlagKey]);
+            Dictionary<EventsCounterKey, EventsCounterValue> expected = new Dictionary<EventsCounterKey, EventsCounterValue>();
+            Assert.Equal(new EventsCounterValue(2, new JValue("value1"), default1),
+                data.Counters[new EventsCounterKey(flag1.Key, flag1.Version, 1)]);
+            Assert.Equal(new EventsCounterValue(1, new JValue("value2"), default1),
+                data.Counters[new EventsCounterKey(flag1.Key, flag1.Version, 2)]);
+            Assert.Equal(new EventsCounterValue(1, new JValue("value99"), default2),
+                data.Counters[new EventsCounterKey(flag2.Key, flag2.Version, 1)]);
+            Assert.Equal(new EventsCounterValue(1, default3, default3),
+                data.Counters[new EventsCounterKey(unknownFlagKey, null, null)]);
         }
     }
